refactor app shell and change folder structure
