--- src/components/appShell/atoms/Logo.tsx
@@ -1,9 +1,0 @@
-const Logo = ({ logoSrc }: { logoSrc: string }) => {
-  return (
-    <>
-      <img src={logoSrc} alt="logo" />
-    </>
-  );
-};
-
-export default Logo;
