+++ src/components/atoms/Logo.tsx
@@ -1,0 +1,7 @@
+export const Logo = ({ logoSrc }: { logoSrc: string }) => {
+  return (
+    <>
+      <img src={logoSrc} alt="logo" />
+    </>
+  );
+};
