+++ src/components/molecules/NavList.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, Story } from '@storybook/react';
 
-import NavList from './NavList';
+import { NavList } from './NavList';
 
 export default {
   component: NavList,
